+++ app/src/main/java/com/enc/music/MainActivity.kt
@@ -1,10 +1,12 @@
 package com.enc.music
 
+import android.content.Intent
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
+import com.enc.music.service.PlaybackService
 import com.enc.music.ui.MusicApp
 import dagger.hilt.android.AndroidEntryPoint
 
@@ -14,6 +16,7 @@ class MainActivity : ComponentActivity() {
         installSplashScreen()
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
+        startService(Intent(this, PlaybackService::class.java))
         setContent {
             MusicApp()
         }