+++ app/src/main/java/com/enc/music/service/PlaybackService.kt
@@ -33,10 +33,7 @@ class PlaybackService : MediaSessionService() {
     }
 
     override fun onDestroy() {
-        mediaSession?.run {
-            player.release()
-            release()
-        }
+        mediaSession?.release()
         mediaSession = null
         super.onDestroy()
     }