+++ app/src/main/java/com/enc/music/ui/navigation/NavHostViewModel.kt
@@ -1,0 +1,11 @@
+package com.enc.music.ui.navigation
+
+import androidx.lifecycle.ViewModel
+import androidx.media3.exoplayer.ExoPlayer
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class NavHostViewModel @Inject constructor(
+    val player: ExoPlayer
+) : ViewModel()